res added some explicit types at some places where event property bag   functions are used for mouse/joystick events to pacify pickier gccs.   Also changed some ints to uints where appropriate.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21499 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/csevent.h
@@ -177,18 +177,18 @@ public:
   csEvent (csEvent const&);
 
   /// Create a mouse event object (old interface)
-  csEvent (csTicks, int type, int x, int y, int button, int modifiers);
+  csEvent (csTicks, int type, int x, int y, uint button, uint32 modifiers);
 
   /// Create a joystick event object (old interface)
-  csEvent (csTicks, int type, int n, int x, int y, uint32 axesChanged, 
-    int button, int modifiers);
+  csEvent (csTicks, int type, uint n, int x, int y, uint32 axesChanged, 
+    uint button, uint32 modifiers);
 
   /// Create an enumerated pointer (mouse/joystick) event object (newer interface)
-  csEvent (csTicks, int type, int n, const int *axes, uint8 numAxes, 
-	   uint32 axesChanged, int button, int modifiers);
+  csEvent (csTicks, int type, uint n, const int32 *axes, uint8 numAxes, 
+	   uint32 axesChanged, uint8 button, uint32 modifiers);
 
   /// Create a command event object
-  csEvent (csTicks, int type, int code, intptr_t info = 0);
+  csEvent (csTicks, int type, uint code, intptr_t info = 0);
 
   /// Destructor
   virtual ~csEvent ();
